# docs/custom-gpt/mind_os_main_openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Notebook Task Manager API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://notebook-task-manager.onrender.com"
        }
    ],
    "paths": {
...
        "/insights": {
            "get": {
                "operationId": "getInsights",
                "summary": "Get Mind OS insights",
                "description": "Read-only insight engine across tasks, projects, memories, and decisions. Use this for patterns, risks, opportunities, strategic follow-up, execution insights, or what the Mind OS notices. Returns dry_run true and writes false. Does not expose freeform insight queries and does not save anything.",
                "parameters": [
                    {
                        "name": "project",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "description": "Optional project filter."
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "minimum": 1,
                            "maximum": 50
                        },
                        "description": "Maximum number of insights and recommended actions to return."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dry-run insights with patterns, risks, opportunities, recommended actions, evidence, signals used, future signals, and warnings."
                    }
                }
            }
        }
    },
...
}